Move service classes to new package
