+++ src/main/java/com/kaustav/launchit/controller/CreditController.java
@@ -1,7 +1,7 @@
 package com.kaustav.launchit.controller;
 
 import com.kaustav.launchit.db.Credit;
-import com.kaustav.launchit.db.CreditService;
+import com.kaustav.launchit.service.CreditService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
