+++ src/main/java/com/kaustav/launchit/controller/InventoryController.java
@@ -1,7 +1,7 @@
 package com.kaustav.launchit.controller;
 
 import com.kaustav.launchit.db.Inventory;
-import com.kaustav.launchit.db.InventoryService;
+import com.kaustav.launchit.service.InventoryService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
