+++ src/main/java/com/kaustav/launchit/controller/OrderController.java
@@ -1,7 +1,7 @@
 package com.kaustav.launchit.controller;
 
 import com.kaustav.launchit.db.OrderEntity;
-import com.kaustav.launchit.db.OrderService;
+import com.kaustav.launchit.service.OrderService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
