+++ src/main/java/com/kaustav/launchit/controller/WarehouseController.java
@@ -1,7 +1,7 @@
 package com.kaustav.launchit.controller;
 
 import com.kaustav.launchit.db.Warehouse;
-import com.kaustav.launchit.db.WarehouseService;
+import com.kaustav.launchit.service.WarehouseService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
