+++ src/main/java/com/kaustav/launchit/service/CreditService.java
@@ -1,4 +1,6 @@
-package com.kaustav.launchit.db;
+package com.kaustav.launchit.service;
+import com.kaustav.launchit.db.Credit;
+import com.kaustav.launchit.db.CreditRepository;
 
 import org.springframework.stereotype.Service;
 import org.slf4j.Logger;
