+++ src/main/java/com/kaustav/launchit/service/OrderItemService.java
@@ -1,4 +1,7 @@
-package com.kaustav.launchit.db;
+package com.kaustav.launchit.service;
+import com.kaustav.launchit.db.OrderItem;
+import com.kaustav.launchit.db.OrderItemId;
+import com.kaustav.launchit.db.OrderItemRepository;
 
 import org.springframework.stereotype.Service;
 import org.slf4j.Logger;
