+++ src/main/java/com/kaustav/launchit/service/RetailerService.java
@@ -1,4 +1,6 @@
-package com.kaustav.launchit.db;
+package com.kaustav.launchit.service;
+import com.kaustav.launchit.db.Retailer;
+import com.kaustav.launchit.db.RetailerRepository;
 
 import org.springframework.stereotype.Service;
 import org.slf4j.Logger;
